feat(ui-common): rename EmptyView and add extraActions prop
