+++ webapp/src/components/App/Singlestudy/explore/Modelization/BindingConstraints/index.tsx
@@ -14,7 +14,7 @@
 
 import { useOutletContext } from "react-router";
 import { StudyMetadata } from "../../../../../../common/types";
-import EmptyView from "../../../../../common/page/SimpleContent";
+import EmptyView from "../../../../../common/page/EmptyView";
 import BindingConstPropsView from "./BindingConstPropsView";
 import {
   getBindingConst,
