+++ webapp/src/components/App/Singlestudy/explore/Modelization/Links/index.tsx
@@ -14,7 +14,7 @@
 
 import { useOutletContext } from "react-router";
 import { StudyMetadata } from "../../../../../../common/types";
-import EmptyView from "../../../../../common/page/SimpleContent";
+import EmptyView from "../../../../../common/page/EmptyView";
 import LinkPropsView from "./LinkPropsView";
 import { getCurrentLink } from "../../../../../../redux/selectors";
 import useAppDispatch from "../../../../../../redux/hooks/useAppDispatch";
