+++ webapp/src/components/common/Matrix/index.tsx
@@ -21,7 +21,7 @@ import { useOutletContext } from "react-router";
 import { StudyMetadata } from "../../../common/types";
 import { MatrixContainer, MatrixHeader, MatrixTitle } from "./styles";
 import MatrixActions from "./components/MatrixActions";
-import EmptyView from "../page/SimpleContent";
+import EmptyView from "../page/EmptyView";
 import { fetchMatrixFn } from "../../App/Singlestudy/explore/Modelization/Areas/Hydro/utils";
 import { AggregateConfig } from "./shared/types";
 import { GridOff } from "@mui/icons-material";
