+++ webapp/src/components/common/page/EmptyView.tsx
@@ -20,10 +20,14 @@ import { SvgIconComponent } from "@mui/icons-material";
 export interface EmptyViewProps {
   title?: string;
   icon?: SvgIconComponent;
+  extraActions?: React.ReactNode;
 }
 
-function EmptyView(props: EmptyViewProps) {
-  const { title, icon: Icon = LiveHelpRoundedIcon } = props;
+function EmptyView({
+  title,
+  icon: Icon = LiveHelpRoundedIcon,
+  extraActions,
+}: EmptyViewProps) {
   const { t } = useTranslation();
 
   return (
@@ -35,8 +39,22 @@ function EmptyView(props: EmptyViewProps) {
         flexDirection: "column",
         alignItems: "center",
         justifyContent: "center",
+        position: "relative",
       }}
     >
+      {extraActions && (
+        <Box
+          sx={{
+            display: "flex",
+            gap: 1,
+            position: "absolute",
+            top: 0,
+            right: 0,
+          }}
+        >
+          {extraActions}
+        </Box>
+      )}
       {Icon && <Icon sx={{ height: 100, width: 100 }} />}
       <div>{title || t("common.noContent")}</div>
     </Box>
